Split tests for ChannelState
The single file testing the state-machine for
the NettingChannelState and NettingChannelEndstate
was very large and was testing a lot of different
state-changes.

In this commit, the file is split up in a separate module,
roughly grouped by the type of state-changes that are
tested
